+++ code/src/sixsq/nuvla/server/resources/email/utils.clj
@@ -1,7 +1,9 @@
 (ns sixsq.nuvla.server.resources.email.utils
   (:require
+    [clojure.java.io :as io]
     [clojure.string :as str]
     [postal.core :as postal]
+    [selmer.parser :as tmpl]
     [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.server.resources.callback :as callback]
     [sixsq.nuvla.server.resources.callback-email-validation :as email-callback]
@@ -10,15 +12,12 @@
     [sixsq.nuvla.server.resources.configuration-nuvla :as config-nuvla]
     [sixsq.nuvla.server.util.response :as r]))
 
+(def base-html (slurp (io/resource "sixsq/nuvla/html-template/base.html")))
 
-(def validation-email-body
-  (partial format
-           (str/join "\n"
-                     ["To validate your email address, visit:"
-                      "\n    %s\n"
-                      "If you did not initiate this request, do NOT click on the link and report"
-                      "this to the service administrator."])))
-
+(def warning-initiate
+  (str/join "\n"
+            ["If you did not initiate this request, do NOT click on the link and report"
+             "this to the service administrator."]))
 
 (def conditions-acceptance
   (partial format
@@ -27,13 +26,47 @@
                       "and Conditions:"
                       "\n    %s\n"])))
 
+(defn validation-email-body
+  [callback-url conditions-url]
+  [:alternative
+   {:type    "text/plain"
+    :content (cond-> (format
+                       (str/join "\n"
+                                 ["To validate your email address, visit:"
+                                  "\n    %s\n"
+                                  warning-initiate])
+                       callback-url)
+                     conditions-url (str (conditions-acceptance conditions-url)))}
+   {:type    "text/html; charset=utf-8"
+    :content (tmpl/render
+               base-html
+               {:title            "Nuvla email validation"
+                :button-text      "Validate"
+                :button-url       callback-url
+                :text-1           "To validate your email address click validate button."
+                :conditions-url   conditions-url
+                :warning-initiate true})}])
 
-(def invitation-email-body
-  (partial format
-           (str/join "\n"
-                     ["You have been invited by \"%s\" to use Nuvla."
-                      "Signup with this email \"%s\" to accept this invite by following this link:"
-                      "\n    %s\n"])))
+
+(defn invitation-email-body
+  [name set-password-url conditions-url]
+  [:alternative
+   {:type    "text/plain"
+    :content (cond-> (format (str/join "\n"
+                                       ["You have been invited by \"%s\" to use Nuvla."
+                                        "To accept the invitation, follow this link:"
+                                        "\n    %s\n"]) name set-password-url)
+                     conditions-url (str (conditions-acceptance conditions-url)))}
+   {:type    "text/html; charset=utf-8"
+    :content (tmpl/render
+               base-html
+               {:title          "Nuvla invitation"
+                :button-text    "Accept invitation"
+                :button-url     set-password-url
+                :text-1         (str
+                                  (format "You have been invited by \"%s\" to use Nuvla. " name)
+                                  "To accept the invitation, click the following button:")
+                :conditions-url conditions-url})}])
 
 
 (defn create-callback [email-id base-uri]
@@ -44,7 +77,8 @@
 
         {{:keys [resource-id]} :body status :status} (crud/add callback-request)]
     (if (= 201 status)
-      (if-let [callback-resource (crud/set-operations (crud/retrieve-by-id-as-admin resource-id) {})]
+      (if-let [callback-resource (crud/set-operations
+                                   (crud/retrieve-by-id-as-admin resource-id) {})]
         (if-let [validate-op (u/get-op callback-resource "execute")]
           (str base-uri validate-op)
           (let [msg "callback does not have execute operation"]
@@ -85,8 +119,7 @@
   (let [{:keys [smtp-username conditions-url]
          :as   nuvla-config} (crud/retrieve-by-id-as-admin config-nuvla/config-instance-url)
 
-        body (cond-> (validation-email-body callback-url)
-                     conditions-url (str (conditions-acceptance conditions-url)))
+        body (validation-email-body callback-url conditions-url)
 
         msg  {:from    (or smtp-username "administrator")
               :to      [address]
@@ -100,55 +133,45 @@
   (let [{:keys [smtp-username, conditions-url]
          :as   nuvla-config} (crud/retrieve-by-id-as-admin config-nuvla/config-instance-url)
 
-        body (cond-> (invitation-email-body (or name id) address set-password-url)
-                     conditions-url (str (conditions-acceptance conditions-url)))
+        body (invitation-email-body (or name id) set-password-url conditions-url)
 
         msg  {:from    (or smtp-username "administrator")
               :to      [address]
-              :subject "email invitation"
+              :subject "Invitation by email for Nuvla service"
               :body    body}]
 
     (send-email nuvla-config msg)))
 
 
-(def password-reset-email-body
-  (partial format
-           (str/join "\n"
-                     ["To reset your password visit:"
-                      "\n    %s\n"
-                      "If you did not initiate this request, do NOT click on the link and report"
-                      "this to the service administrator."])))
-
-
-(defn send-password-reset-email [callback-url address]
-  (let [{:keys [smtp-username] :as nuvla-config} (crud/retrieve-by-id-as-admin config-nuvla/config-instance-url)
-
-        body (password-reset-email-body callback-url)
-
-        msg  {:from    (or smtp-username "administrator")
-              :to      [address]
-              :subject "reset password"
-              :body    body}]
-
-    (send-email nuvla-config msg)))
-
-(def password-set-email-body
-  (partial format
-           (str/join "\n"
-                     ["To set your password visit:"
-                      "\n    %s\n"
-                      "If you did not initiate this request, do NOT click on the link and report"
-                      "this to the service administrator."])))
+(defn password-set-email-body
+  [set-password-url]
+  [:alternative
+   {:type    "text/plain"
+    :content (format
+               (str/join "\n"
+                         ["To set your password visit:"
+                          "\n    %s\n"
+                          warning-initiate])
+               set-password-url)}
+   {:type    "text/html; charset=utf-8"
+    :content (tmpl/render
+               base-html
+               {:title            "Nuvla set password"
+                :button-text      "Set new password"
+                :button-url       set-password-url
+                :text-1           "To set your password click following button:"
+                :warning-initiate true})}])
 
 
 (defn send-password-set-email [set-password-url address]
-  (let [{:keys [smtp-username] :as nuvla-config} (crud/retrieve-by-id-as-admin config-nuvla/config-instance-url)
+  (let [{:keys [smtp-username] :as nuvla-config} (crud/retrieve-by-id-as-admin
+                                                   config-nuvla/config-instance-url)
 
         body (password-set-email-body set-password-url)
 
         msg  {:from    (or smtp-username "administrator")
               :to      [address]
-              :subject "set password"
+              :subject "Set password for Nuvla service"
               :body    body}]
 
     (send-email nuvla-config msg)))