+++ code/test/sixsq/nuvla/server/resources/hook_reset_password_lifecycle_test.clj
@@ -39,7 +39,9 @@
 
                   ;; WARNING: This is a fragile!  Regex matching to recover callback URL.
                   postal/send-message (fn [_ {:keys [body]}]
-                                        (let [url (second (re-matches #"(?s).*visit:\n\n\s+(.*?)\n.*" body))]
+                                        (let [url (->> body second :content
+                                                       (re-matches #"(?s).*visit:\n\n\s+(.*?)\n.*")
+                                                       second)]
                                           (reset! reset-link url))
                                         {:code 0, :error :SUCCESS, :message "OK"})]