+++ code/test/sixsq/nuvla/server/resources/session_password_lifecycle_test.clj
@@ -50,7 +50,9 @@
 
                   ;; WARNING: This is a fragile!  Regex matching to recover callback URL.
                   postal/send-message (fn [_ {:keys [body]}]
-                                        (let [url (second (re-matches #"(?s).*visit:\n\n\s+(.*?)\n.*" body))]
+                                        (let [url (->> body second :content
+                                                       (re-matches #"(?s).*visit:\n\n\s+(.*?)\n.*")
+                                                       second)]
                                           (reset! validation-link url))
                                         {:code 0, :error :SUCCESS, :message "OK"})]